# openapi-docs.yaml
openapi: 3.0.1
info:
  title: Price Negotiation App
  contact:
    name: Łukasz Górski
    url: https://www.linkedin.com/in/lukasz-gorski-lukegor/
    email: lukaszgorski02@gmail.com
  license:
    name: Apache License 2.0
    url: https://opensource.org/license/apache-2-0/
  version: v1
paths:
...
  /api/v1/Negotiations/Negotiation/all:
    get:
      tags:
        - Negotiation
      summary: Retrieves a list of all negotiations.
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Negotiation'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Negotiation'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Negotiation'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/v1/Negotiations/Negotiation/{id}':
    get:
      tags:
        - Negotiation
      summary: Retrieves a specific negotiation by its unique identifier.
      parameters:
        - name: id
          in: path
          description: The unique identifier of the negotiation to retrieve.
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Negotiation'
            application/json:
              schema:
                $ref: '#/components/schemas/Negotiation'
            text/json:
              schema:
                $ref: '#/components/schemas/Negotiation'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags:
        - Negotiation
      summary: Updates a specific negotiation by its unique identifier.
      parameters:
        - name: id
          in: path
          description: The unique identifier of the negotiation to update.
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: The updated negotiation data.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Negotiation'
          text/json:
            schema:
              $ref: '#/components/schemas/Negotiation'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Negotiation'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Server Error
    delete:
      tags:
        - Negotiation
      summary: Deletes a specific negotiation by its unique identifier.
      parameters:
        - name: id
          in: path
          description: The unique identifier of the negotiation to delete.
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /api/v1/Negotiations/Negotiation/response:
    patch:
      tags:
        - Negotiation
      summary: Responds to a negotiation proposal.
      parameters:
        - name: negotiationId
          in: query
          description: The unique identifier of the negotiation to update.
          schema:
            type: integer
            format: int32
        - name: isApproved
          in: query
          description: A flag indicating whether the proposal is approved or not.
          schema:
            type: boolean
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Server Error
...
components:
  schemas:
...
    Negotiation:
      required:
        - productId
        - proposedPrice
        - retriesLeft
        - status
        - userId
      type: object
      properties:
        id:
          type: integer
          format: int32
        productId:
          minLength: 1
          type: string
        proposedPrice:
          type: number
          format: double
        isAccepted:
          type: boolean
          nullable: true
        retriesLeft:
          maximum: 2
          minimum: 0
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/NegotiationStatus'
        userId:
          minLength: 1
          type: string
      additionalProperties: false
...
    NegotiationStatus:
      enum:
        - 0
        - 1
        - 2
      type: integer
      format: int32
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: Authorization header using the Bearer scheme for JWT
      name: Authorization
      in: header
security:
  - Bearer: [ ]